+++ src/app/mobile/details-dialog.tsx
@@ -1,0 +1,1 @@
+export function DetailsDialog() {}
